refactor(server): Minimize code in the main module
