+++ crates/server/src/main.rs
@@ -1,18 +1,8 @@
-use std::{io, net::SocketAddr, str::FromStr, sync::Arc, time::Duration};
+use std::{net::SocketAddr, sync::Arc, time::Duration};
 
-use apalis::{
-    cron::{CronStream, Schedule},
-    layers::{
-        retry::{RetryLayer, RetryPolicy},
-        tracing::TraceLayer,
-        TimeoutLayer,
-    },
-    prelude::*,
-    sqlite::SqliteStorage,
-};
 use axum::{http::HeaderValue, routing::get, Router};
 use sqlx::{migrate::MigrateDatabase, Sqlite, SqlitePool};
-use tokio::{signal, sync::broadcast};
+use tokio::sync::broadcast;
 use tower_http::{
     cors::{Any, CorsLayer},
     services::{ServeDir, ServeFile},
@@ -20,20 +10,19 @@ use tower_http::{
     trace::{DefaultMakeSpan, DefaultOnRequest, DefaultOnResponse, TraceLayer as HttpTraceLayer},
     LatencyUnit,
 };
-use tracing::{debug, error, info, Level};
+use tracing::{error, info, Level};
 use ws::ws_handler;
 
-use crate::{
-    config::env_config, job::Notification, models::service::Service, routes::routes,
-    ws::Event as WsEvent,
-};
+use crate::{config::env_config, routes::routes, ws::Event as WsEvent};
 
 mod auth;
 mod config;
 mod job;
 mod models;
+mod monitors;
 mod routes;
 mod service;
+mod utils;
 mod ws;
 
 type AppState = Arc<AppStateInner>;
@@ -112,10 +101,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         info!("listening on port {}", env.port);
         axum::serve(
             listener,
+            // this is required for ws to work, I don't know why.
             app.into_make_service_with_connect_info::<SocketAddr>(),
         )
         .with_graceful_shutdown(async {
-            shutdown_signal()
+            utils::shutdown_signal()
                 .await
                 .expect("failed to install Ctrl+C handler");
             info!("Ctrl+C Received, Shutting down");
@@ -125,75 +115,14 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     };
 
     // run background workers
-    let monitors = async {
-        let notification_storage: SqliteStorage<Notification> =
-            SqliteStorage::new(state.pool.clone());
-
-        let monitor_storage: SqliteStorage<Service> = SqliteStorage::new(state.pool.clone());
-
-        let schedule = Schedule::from_str("* * * * * *")?;
-        let cron_timer = WorkerBuilder::new("uptime-timer")
-            .data(service::TimerService::new(state.pool.clone()))
-            .stream(CronStream::new(schedule).into_stream())
-            // curtesy of https://github.com/geofmureithi/apalis/issues/297
-            .chain(|s| {
-                s.layer(TimeoutLayer::new(Duration::from_secs(60)))
-                    .layer(TraceLayer::new())
-            })
-            .build_fn(service::run_timer_cron_service);
-
-        let notify_worker = WorkerBuilder::new("notification-worker")
-            .layer(TraceLayer::new())
-            .with_storage(notification_storage.clone())
-            .build_fn(job::notify);
-
-        let monitor_worker = WorkerBuilder::new("monitor-worker")
-            .data(state.clone())
-            .layer(RetryLayer::new(RetryPolicy::retries(3)))
-            .layer(TraceLayer::new())
-            .with_storage(monitor_storage)
-            .build_fn(job::job_monitor);
-
-        Monitor::<TokioExecutor>::new()
-            .register_with_count(1, cron_timer)
-            .register_with_count(1, notify_worker)
-            .register_with_count(2, monitor_worker)
-            .shutdown_timeout(Duration::from_secs(10))
-            .on_event(|e| {
-                let worker_id = e.id();
-                match e.inner() {
-                    Event::Start => {
-                        info!(target: "worker", worker = %worker_id, "started");
-                    }
-                    Event::Engage => {
-                        debug!(target: "worker", worker = %worker_id, "engaged");
-                    }
-                    Event::Idle => {
-                        debug!(target: "worker", worker = %worker_id, "idle");
-                    }
-                    Event::Error(e) => {
-                        error!(target: "worker", worker = %worker_id, "error: {e:?}");
-                    }
-                    Event::Exit => {
-                        info!(target: "worker", worker = %worker_id, "exited");
-                    }
-                    Event::Stop => {
-                        info!(target: "worker", worker = %worker_id, "stopped");
-                    }
-                }
-            })
-            .run_with_signal(async {
-                shutdown_signal().await?;
-                info!(target: "worker","Ctrl+C Received, Shutting down");
-                Ok(())
-            })
-            .await?;
+    let workers = async {
+        monitors::monitors(&state).await?;
         info!("workers stopped");
         Ok::<(), Box<dyn std::error::Error>>(())
     };
 
     // Start axum server and workers
-    match tokio::join!(http, monitors) {
+    match tokio::join!(http, workers) {
         (Err(e), _) => error!("Server exited with error: {e}"),
         (_, Err(e)) => error!("Worker exited with error: {e}"),
         _ => (),
@@ -201,24 +130,4 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     info!("closing db connection");
     state.pool.close().await;
     Ok(())
-}
-
-async fn shutdown_signal() -> io::Result<()> {
-    let ctrl_c = signal::ctrl_c();
-
-    #[cfg(unix)]
-    let terminate = async {
-        signal::unix::signal(signal::unix::SignalKind::terminate())
-            .expect("failed to install signal handler")
-            .recv()
-            .await;
-    };
-
-    #[cfg(not(unix))]
-    let terminate = std::future::pending::<()>();
-
-    tokio::select! {
-        v = ctrl_c => v,
-        _ = terminate => Ok(()),
-    }
 }
